+++ slick_reporting/__init__.py
@@ -1,6 +1,6 @@
 
 default_app_config = 'slick_reporting.apps.ReportAppConfig'
 
-VERSION = (0, 5, 1)
+VERSION = (0, 5, 2)
 
-__version__ = '0.5.1'
+__version__ = '0.5.2'