+++ slick_reporting/generator.py
@@ -230,7 +230,6 @@ class ReportGenerator(object):
                 # group_by_filter = self.kwargs_filters.get(self.group_by, '')
                 # qs = self.group_by_field.related_model.objects
                 # if group_by_filter:
-                #     import pdb; pdb.set_trace()
                 #     lookup = 'pk__in' if isinstance(group_by_filter, Iterable) else 'pk'
                 #     qs = qs.filter(**{lookup: group_by_filter})
                 # self.main_queryset = qs.values()
@@ -300,11 +299,8 @@ class ReportGenerator(object):
                     fields_on_report = [x for x in window_cols if x['ref'] in dependencies_names]
                     for field in fields_on_report:
                         self._report_fields_dependencies[window][field['name']] = col_data['name']
-            # import pdb; pdb.set_trace()
             for col_data in window_cols:
                 klass = col_data['ref']
-                # if getattr(klass, 'name', '') not in klasses_names:
-                #     continue
                 name = col_data['name']
 
                 # if column has a dependency then skip it
@@ -519,7 +515,7 @@ class ReportGenerator(object):
         cols = self.time_series_columns or []
         series = self._get_time_series_dates(self.time_series_pattern)
 
-        for dt in series:
+        for index, dt in enumerate(series):
             for col in cols:
                 magic_field_class = None
 
@@ -531,7 +527,7 @@ class ReportGenerator(object):
                 _values.append({
                     'name': magic_field_class.name + 'TS' + dt[1].strftime('%Y%m%d'),
                     'original_name': magic_field_class.name,
-                    'verbose_name': self.get_time_series_field_verbose_name(magic_field_class, dt),
+                    'verbose_name': self.get_time_series_field_verbose_name(magic_field_class, dt, index, series),
                     'ref': magic_field_class,
                     'start_date': dt[0],
                     'end_date': dt[1],
@@ -540,7 +536,7 @@ class ReportGenerator(object):
                 })
         return _values
 
-    def get_time_series_field_verbose_name(self, computation_class, date_period):
+    def get_time_series_field_verbose_name(self, computation_class, date_period, index, series, pattern=None):
         """
         Sent the column data to construct a verbose name.
         Default implementation is delegated to the ReportField.get_time_series_field_verbose_name
@@ -550,7 +546,9 @@ class ReportGenerator(object):
         :param date_period: a tuple of (start_date, end_date)
         :return: a verbose string
         """
-        return computation_class.get_time_series_field_verbose_name(date_period)
+        pattern = pattern or self.time_series_pattern
+        return computation_class.get_time_series_field_verbose_name(date_period, index, series,
+                                                                    pattern)
 
     def get_custom_time_series_dates(self):
         """